move k2_util.h to include folder to avoid relative header file path
